refactor: split oversized engine files and add unit tests
Split engine/mod.rs (645→154 lines), engine/container.rs (703→274 lines),
and engine/volume.rs (557→272 lines) — all were over the 500-line hard limit.

New modules:
- engine/lifecycle.rs  — up, down, restart commands
- engine/query.rs      — ps, logs, exec, pull, remove_orphans, attach_logs
- engine/container_config.rs — restart policy, log config, healthcheck, resources, ulimits
- engine/container_misc.rs   — device, blkio, tmpfs, label-file, opt_vec/opt_map
- engine/volume_mounts.rs    — build_binds, build_mounts helpers

Unit tests added for all pure (non-Podman) helpers in:
engine/container_config, engine/container_misc, engine/network, engine/profiles,
engine/volume_mounts. Coverage rises from 41% to 50%.

Enable coverage-threshold: 50 in ci.yml (gate was disabled). Integration tests
for Podman-touching engine code are tracked in #51.
