+++ .github/workflows/ci.yml
@@ -14,3 +14,4 @@ jobs:
     uses: Glyndor/.github/.github/workflows/rust-ci.yml@bc62ab8f95465df2b2f6f23775305f10a44e1502 # main
     with:
       extra-test-os: '["macos-latest", "windows-latest"]'
+      coverage-threshold: 50
